Isolated supabase config functions in config.py
